SplitPane.dividerSize => 1

+++ j/src/org/armedbear/j/DefaultLookAndFeel.java
@@ -2,7 +2,7 @@
  * DefaultLookAndFeel.java
  *
  * Copyright (C) 2000-2003 Peter Graves
- * $Id: DefaultLookAndFeel.java,v 1.4 2003-07-24 16:27:08 piso Exp $
+ * $Id: DefaultLookAndFeel.java,v 1.5 2003-07-24 19:41:20 piso Exp $
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -112,7 +112,7 @@ public final class DefaultLookAndFeel extends DefaultMetalTheme
         table.put("SplitPaneUI", "javax.swing.plaf.basic.BasicSplitPaneUI");
         table.put("ScrollBarUI", "javax.swing.plaf.basic.BasicScrollBarUI");
         table.put("TreeUI", "javax.swing.plaf.basic.BasicTreeUI");
-        table.put("SplitPane.dividerSize", new Integer(3));
+        table.put("SplitPane.dividerSize", new Integer(1));
         table.put("ScrollBar.background", new Color(0xe0e0e0));
         table.put("ScrollBar.foreground", new Color(0xc0c0c0));
         table.put("ScrollBar.track", new Color(0xe0e0e0));
